+++ src/bin/dolos/main.rs
@@ -143,6 +143,7 @@ pub struct Config {
     pub retries: Option<gasket::retries::Policy>,
     pub mithril: Option<MithrilConfig>,
     pub snapshot: Option<SnapshotConfig>,
+    pub chain: Option<dolos::adapters::ChainConfig>,
 
     #[serde(default)]
     pub logging: LoggingConfig,