+++ src/mempool.rs
@@ -283,9 +283,9 @@ impl MempoolStore for Mempool {
         {
             let report = self.evaluate(&tx)?;
 
-            for eval in report.iter() {
+            for eval in report {
                 if !eval.success {
-                    return Err(MempoolError::Phase2ExplicitError);
+                    return Err(MempoolError::Phase2ExplicitError(eval.logs.clone()));
                 }
             }
         }